Refactor: make smaller more manageable components for check form (feature chain branch: 2 of 4) (#723)
* feat: extract multihttp form layout to own file

* fix: toggle requests section in multihttp test

* fix: remove type cast from multihttp components

* Refactor: make smaller more manageable components for check form (feature chain branch: 3 of 4) (#724)

* feat: add scripted layout

* feat: add ping layout

* fix: fix layouts

* feat: added dnscheck layout

* feat: add traceroute layout

* Refactor: make smaller more manageable components for check form (feature chain branch: 4 of 4) (#725)

* feat: move state management to collapse

* feat: move checkformalerts to layout files
