+++ src/components/CheckForm/FormLayouts/CheckPingLayout.tsx
@@ -1,0 +1,50 @@
+import React from 'react';
+import { GrafanaTheme2 } from '@grafana/data';
+import { useStyles2 } from '@grafana/ui';
+import { css } from '@emotion/css';
+
+import { CheckFormValuesPing, CheckType } from 'types';
+import { CheckEnabled } from 'components/CheckEditor/FormComponents/CheckEnabled';
+import { CheckIpVersion } from 'components/CheckEditor/FormComponents/CheckIpVersion';
+import { CheckJobName } from 'components/CheckEditor/FormComponents/CheckJobName';
+import { CheckPublishedAdvanceMetrics } from 'components/CheckEditor/FormComponents/CheckPublishedAdvanceMetrics';
+import { CheckTarget } from 'components/CheckEditor/FormComponents/CheckTarget';
+import { PingCheckFragment } from 'components/CheckEditor/FormComponents/PingCheckFragment';
+import { ProbeOptions } from 'components/CheckEditor/ProbeOptions';
+import { CheckFormAlert } from 'components/CheckFormAlert';
+import { CheckUsage } from 'components/CheckUsage';
+import { Collapse } from 'components/Collapse';
+import { LabelField } from 'components/LabelField';
+
+export const CheckPingLayout = () => {
+  const styles = useStyles2(getStyles);
+
+  return (
+    <>
+      <Collapse label="General settings" isOpen>
+        <CheckEnabled />
+        <CheckJobName />
+        <CheckTarget checkType={CheckType.PING} />
+        <ProbeOptions checkType={CheckType.PING} />
+        <CheckPublishedAdvanceMetrics />
+        <CheckUsage />
+      </Collapse>
+      <Collapse label="Advanced options">
+        <div className={styles.maxWidth}>
+          <LabelField<CheckFormValuesPing> />
+          <CheckIpVersion checkType={CheckType.PING} name="settings.ping.ipVersion" />
+          <PingCheckFragment />
+        </div>
+      </Collapse>
+      <Collapse label="Alerting">
+        <CheckFormAlert />
+      </Collapse>
+    </>
+  );
+};
+
+const getStyles = (theme: GrafanaTheme2) => ({
+  maxWidth: css({
+    maxWidth: `500px`,
+  }),
+});
