+++ src/components/CheckFormAlert.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Controller, useFormContext } from 'react-hook-form';
 import { GrafanaTheme2 } from '@grafana/data';
 import { Field, Select, useStyles2 } from '@grafana/ui';
@@ -6,20 +6,9 @@ import { css } from '@emotion/css';
 
 import { CheckFormValues } from 'types';
 
-import { Collapse } from './Collapse';
 import { ALERT_SENSITIVITY_OPTIONS } from './constants';
 
-const getStyles = (theme: GrafanaTheme2) => ({
-  marginBottom: css`
-    margin-bottom: ${theme.spacing(2)};
-  `,
-  link: css`
-    text-decoration: underline;
-  `,
-});
-
 export const CheckFormAlert = () => {
-  const [showAlerting, setShowAlerting] = useState(false);
   const styles = useStyles2(getStyles);
   const { watch } = useFormContext<CheckFormValues>();
   const alertSensitivity = watch('alertSensitivity');
@@ -27,7 +16,7 @@ export const CheckFormAlert = () => {
   const isCustomSensitivity = !Boolean(ALERT_SENSITIVITY_OPTIONS.find((option) => option.value === alertSensitivity));
 
   return (
-    <Collapse label="Alerting" onToggle={() => setShowAlerting(!showAlerting)} isOpen={showAlerting}>
+    <>
       <div className={styles.marginBottom}>
         <p>
           Synthetic Monitoring provides some default alert rules via Cloud Alerting. By selecting an alert sensitivity,
@@ -68,6 +57,15 @@ export const CheckFormAlert = () => {
           }}
         />
       </Field>
-    </Collapse>
+    </>
   );
 };
+
+const getStyles = (theme: GrafanaTheme2) => ({
+  marginBottom: css({
+    marginBottom: theme.spacing(2),
+  }),
+  link: css({
+    textDecoration: `underline`,
+  }),
+});
