+++ src/components/TLSConfig.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { DeepMap, FieldError, useFormContext } from 'react-hook-form';
 import { OrgRole } from '@grafana/data';
 import { Container, Field, Input, TextArea } from '@grafana/ui';
@@ -6,7 +6,6 @@ import { Container, Field, Input, TextArea } from '@grafana/ui';
 import { CheckFormValuesHttp, CheckFormValuesTcp, CheckType } from 'types';
 import { hasRole } from 'utils';
 import { validateTLSCACert, validateTLSClientCert, validateTLSClientKey, validateTLSServerName } from 'validation';
-import { Collapse } from 'components/Collapse';
 import { HorizontalCheckboxField } from 'components/HorizonalCheckboxField';
 
 interface Props {
@@ -14,7 +13,6 @@ interface Props {
 }
 
 export const TLSConfig = ({ checkType }: Props) => {
-  const [showTLS, setShowTLS] = useState(false);
   const {
     register,
     formState: { errors },
@@ -23,7 +21,7 @@ export const TLSConfig = ({ checkType }: Props) => {
   const errs = isErrorsHttp(errors) ? errors.settings?.http : isErrorsTcp(errors) ? errors.settings?.tcp : undefined;
 
   return (
-    <Collapse label="TLS config" onToggle={() => setShowTLS(!showTLS)} isOpen={showTLS}>
+    <>
       <HorizontalCheckboxField
         id="tls-config-skip-validation"
         disabled={!isEditor}
@@ -109,7 +107,7 @@ export const TLSConfig = ({ checkType }: Props) => {
           />
         </Field>
       </Container>
-    </Collapse>
+    </>
   );
 };
