--- src/components/TcpSettings/index.ts
@@ -1,1 +1,0 @@
-export { TcpSettingsForm } from './TcpSettings';
